@@ -1,13 +1,20 @@
 module github.com/dakusui/jqplusplus
 
-go 1.22.2
+go 1.23.0
+
+toolchain go1.24.5
 
 require (
+	github.com/BurntSushi/toml v1.6.0
 	github.com/itchyny/gojq v0.12.17
+	github.com/titanous/json5 v1.0.0
+	gopkg.in/yaml.v3 v3.0.1
 	sigs.k8s.io/yaml v1.6.0
 )
 
 require (
+	github.com/google/go-cmp v0.6.0 // indirect
 	github.com/itchyny/timefmt-go v0.1.6 // indirect
 	go.yaml.in/yaml/v2 v2.4.2 // indirect
+	golang.org/x/text v0.25.0 // indirect
 )
